feat: Add user role check for editing article
The code changes in `ArticleDetails.js` add a user role check to the rendering of the "Edit Article" button. Now, the button will only be displayed if the user is logged in and has the role of "Admin". This enhancement improves the security and user experience of the application by restricting access to the edit functionality to authorized users.

Co-authored-by: Yael Gluskin