feat: Add LogOut component and route
This commit adds the LogOut component and route to the application. The LogOut component handles the functionality of logging out the user by clearing the user context and navigating to the home page. This feature enhances the user experience by providing a way for users to securely log out of the application.

Co-authored-by: Yael Gluskin