feat: Add CreateUserForm component
This commit adds the CreateUserForm component, which provides a form for creating new users in the admin area. The form includes fields for username, password, email, and user role. Upon submission, the form sends a POST request to the backend API to create a new user. This enhancement improves the functionality of the application by allowing administrators to easily create new users.